requirements.txt

appdirs==1.4.3
arrow==0.10.0
backports-abc==0.5
certifi==2017.4.17
chardet==3.0.4
idna==2.5
praw==5.0.0
prawcore==0.11.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.18.1
lowScore==6.0.7
six==1.10.0
tornado==4.5.1
update-checker==0.16
urllib3==1.21.1

